Sprint_9 Настроить работу http сервера и проверить через созданного клиента в тестах V1
