+++ src/adapters/DurationAdapter.java
@@ -1,4 +1,4 @@
-package Adapters;
+package adapters;
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
