+++ src/http/handlers/BaseHttpHandler.java
@@ -1,4 +1,4 @@
-package API.Handlers;
+package http.handlers;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import service.TaskManager;
